♻️ refactor(heterogeneous-agent): move StatusGuide out of features/Electron
StatusGuide renders on web too, so it doesn't belong under features/Electron.
Seams the single electron system-service call behind a openExternalLink
platform twin so the web micro-app builds never pull in the electron layer.
