+++ src/features/HeterogeneousAgent/StatusGuide/openExternalLink.desktop.ts
@@ -1,0 +1,5 @@
+import { electronSystemService } from '@/services/electron/system';
+
+export const openExternalLink = async (url: string): Promise<void> => {
+  await electronSystemService.openExternalLink(url);
+};
