+++ src/features/HeterogeneousAgent/StatusGuide/openExternalLink.ts
@@ -1,0 +1,3 @@
+export const openExternalLink = async (url: string): Promise<void> => {
+  window.open(url, '_blank', 'noopener,noreferrer');
+};
